atualização

+++ Python 2024/aula3/exercicio4.py
@@ -3,7 +3,7 @@
 #o guidão tem /2 do aro. manete 1/4 do aro da roda. o quadro será o mesmo do aro 
 
 print("vamos juntar informação para a sua bike?")
-aro = int(input("qual o tamanho do aro da sua Bike?? "))
+aro = float(input("qual o tamanho do aro da sua Bike?? "))
 guidão = aro/2
 manete = aro/4
 quadrobike = aro
